Fixes for a few warnings on 64-bit
git-svn-id: https://liblo.svn.sourceforge.net/svnroot/liblo/trunk@83 8a0476ba-7714-0410-bd95-fb3ea4a72f84

+++ src/message.c
@@ -410,7 +410,7 @@ void lo_arg_pp_internal(lo_type type, void *data, int bigendian)
 	break;
 
     case LO_INT64:
-	printf("%lld", val64.i);
+	printf("%lld", (long long int)val64.i);
 	break;
     
     case LO_TIMETAG:
